refactor(taxonomic-filter): rename recentPinnedTaxonomicGroupsLogic.ts to recentPinnedTabDefinitions.ts
The file no longer contains a kea logic — just exports a frozen
constant. Convergent qa-swarm finding (qa-team + paul + xp) on the
prior iteration: the *Logic suffix misleads a future reader grepping
for kea actions/selectors. Match the filename to the contents.

Two import sites updated.

Generated-By: PostHog Code
Task-Id: 2649f7ae-c1f7-40ae-8866-be024f3f1285
